# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/shelves": {
            "get": {
                "tags": [
                    "Shelf"
                ],
                "summary": "Lấy danh sách các kệ hàng",
                "operationId": "d0ea09eefcb4a18d9f7f3ce7643119d5",
                "responses": {
                    "200": {
                        "description": "Danh sách các kệ hàng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 7
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Kệ 1"
                                            },
                                            "warehouse_name": {
                                                "type": "string",
                                                "example": "Kho nguyên vật liệu 1"
                                            },
                                            "category_name": {
                                                "type": "string",
                                                "example": "Nhựa HDPE"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-10-10T14:13:35.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": null,
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hiện tại không có kệ hàng nào",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lỗi khi lấy danh sách kệ hàng"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Hiện tại không có kệ hàng nào."
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi khi lấy danh sách kệ hàng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lỗi khi lấy danh sách kệ hàng"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Chi tiết lỗi"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/shelf/add": {
            "post": {
                "tags": [
                    "Shelf"
                ],
                "summary": "Tạo một kệ hàng mới",
                "operationId": "b0f3beb0f7a8ec535b29f76f45871b19",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Test thêm kệ hàng"
                                    },
                                    "warehouse_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "number_of_levels": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "storage_capacity": {
                                        "type": "integer",
                                        "example": 1000
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo kệ hàng thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tạo kệ hàng thành công"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 11
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Test thêm kệ hàng"
                                                },
                                                "warehouse_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "number_of_levels": {
                                                    "type": "integer",
                                                    "example": 8
                                                },
                                                "storage_capacity": {
                                                    "type": "integer",
                                                    "example": 1000
                                                },
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-11T08:58:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-11T08:58:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Tạo kệ hàng thất bại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tạo kệ hàng thất bại"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Chi tiết lỗi"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/shelf/{id}": {
            "get": {
                "tags": [
                    "Shelf"
                ],
                "summary": "Lấy thông tin chi tiết của một kệ hàng",
                "operationId": "e4c27e1b8c3a5ea78b0aac68dc3e13cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của kệ hàng",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin chi tiết của kệ hàng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Kệ 1"
                                        },
                                        "warehouse_name": {
                                            "type": "string",
                                            "example": "Kho thành phẩm 1"
                                        },
                                        "category_name": {
                                            "type": "string",
                                            "example": "Chai nhựa HDPE"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-10-11T08:42:38.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy kệ hàng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lỗi khi lấy thông tin kệ hàng"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Không tìm thấy kệ hàng với ID: 1111"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi khi lấy thông tin kệ hàng",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lỗi khi lấy thông tin kệ hàng"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Chi tiết lỗi"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}